refactor: collapse update deployment boundary
