+++ apps/server/vibesensor/use_cases/updates/workflow_executor.py
@@ -4,7 +4,9 @@ from __future__ import annotations
 
 from vibesensor.use_cases.updates.firmware import FirmwareRefresher
 from vibesensor.use_cases.updates.models import UpdateExecutionOutcome
-from vibesensor.use_cases.updates.release_deployment import UpdateReleaseDeployer
+from vibesensor.use_cases.updates.release_deployment import (
+    UpdateReleaseDeploymentCoordinator,
+)
 from vibesensor.use_cases.updates.release_staging import ServerReleaseStager
 from vibesensor.use_cases.updates.restart_scheduler import UpdateRestartScheduler
 from vibesensor.use_cases.updates.run_models import (
@@ -23,7 +25,7 @@ class UpdateWorkflowExecutor:
     """Execute a prepared release plan while delegating side effects to focused collaborators."""
 
     __slots__ = (
-        "_deployer",
+        "_deployment",
         "_firmware_refresher",
         "_restart_scheduler",
         "_stager",
@@ -35,14 +37,14 @@ class UpdateWorkflowExecutor:
         self,
         *,
         stager: ServerReleaseStager,
-        deployer: UpdateReleaseDeployer,
+        deployment: UpdateReleaseDeploymentCoordinator,
         firmware_refresher: FirmwareRefresher,
         restart_scheduler: UpdateRestartScheduler,
         status: UpdateStatusTracker,
         transport_coordinator: UpdateTransportCoordinator,
     ) -> None:
         self._stager = stager
-        self._deployer = deployer
+        self._deployment = deployment
         self._firmware_refresher = firmware_refresher
         self._restart_scheduler = restart_scheduler
         self._status = status
@@ -83,7 +85,7 @@ class UpdateWorkflowExecutor:
         plan: InstallServerReleasePlan,
     ) -> UpdateExecutionOutcome:
         async with self._stager.stage(plan.release) as staged_release:
-            await self._deployer.deploy(staged_release)
+            await self._deployment.deploy(staged_release)
         await self._complete_success(
             workflow.prepared.transport_session,
             message="Update completed successfully",
